Optimize pattern matching (#6)
* chore: Improve release workflow and enable version embedding

* feat: add cpu profiling

* refactor: remove old profiling cmd

* perf: optimize pattern matching

* refactor: move optimization to Matcher

* feat(pattern): Optimize pattern matching

This commit optimizes the pattern matching by adding a fast path for the `hardLiteral`.
If the `hardLiteral` (the longest literal in the pattern) is not found in the input, the pattern matching is skipped.
This avoids unnecessary calls to `bytes.Index` and improves the performance of the `Test` function.

* refactor(pattern): Rename hardLiteral to longestLiteral

This commit renames the `hardLiteral` field to `longestLiteral` to better
reflect its purpose. This field holds the longest literal in the pattern
and is used to quickly discard non-matching lines.